Refactor app layout to fix auth/cache issues Routes were not refreshing and re-fetching user data as expected to check if logged in.
Navbar had admin-only items, but it was not rendering correctly. This was because the navbar shared a common layout and was not re-rendered:
https://beta.nextjs.org/docs/api-reference/file-conventions/layout#good-to-know
"When navigating from /dashboard to /dashboard/settings, dashboard/settings/page.tsx will be rendered on the server because it the segment
that changed, while dashboard/layout.tsx will not be re-rendered because it is a common layout between the two routes."

The docs appear to be incorrect, as calling router.refresh() after log out did not seem to refresh the navbar.
Or maybe it did but only for the "current route", and not the whole route group, so the navbar was inconsistent between pages.
https://beta.nextjs.org/docs/routing/linking-and-navigating#invalidating-the-cache
"The cache can be invalidated using router.refresh(). For more information, see the API reference."
https://beta.nextjs.org/docs/data-fetching/mutating
"After a data mutation, you can use router.refresh() to refresh (fetch updated data and re-render on the server) the current route from the root layout down."

The docs appear to be incorrect, or at least a full page load does not appear to mean the root layout will re-fetch data and re-render on the server:
https://beta.nextjs.org/docs/routing/defining-routes#good-to-know
"Navigating across multiple root layouts will cause a full page load (as opposed to a client-side navigation).
For example, navigating from /cart that uses app/(shop)/layout.js to /blog that uses app/(marketing)/layout.js will cause a full page load.
This only applies to multiple root layouts."

I tried to access the url/pathname from the root layout to conditionally render layouts but that did not work: https://github.com/vercel/next.js/issues/43704

Eventually the fix was:
- move login page to root /app directory, above either route group. Possibly has no effect, but trying to cross multiple routes to ensure data refresh.
- login page checks user cookie and redirects to /training-studio if cookie exists
- set dynamic = 'force-dynamic' in login page to force dynamic rendering and dynamic data fetching by disabling all caching of fetch requests and always revalidating.
- make the root layout minimal to wrap the route group layouts
- add custom layouts to each route group
- the training-app route group layout checks user cookie, redirects to /login if none
- as a fallback the layout renders JSX conditionally on user, but that's possibly redundant
- training-app layout passes isAdmin boolean to its navbar
- log out button calls fetch('/logout'), which is caught by middleware, which deletes cookie
- log out button calls router.push('/').  Redirect in middleware did not work for this
- the link to /training-studio on homepage (which middleware can redirect to /login), has prefetch set to false
- with prefetch set to true, user was able to click back to cached /training-studio after logging out